+++ CHANGELOG.md
@@ -10,7 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 > Use [this search for a list of all CHANGELOG.md files in this repo](https://github.com/search?q=repo%3Aopen-telemetry%2Fopentelemetry-python-contrib+path%3A**%2FCHANGELOG.md&type=code).
 
 ## Unreleased
-
+### Changed
+- `opentelemetry-instrumentation`: Batch all packages into a single `pip install` call in `bootstrap.py` to improve performance
+  ([#4530](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4530))
+  
 ### Added
 
 - Add `BaggageLogProcessor` to `opentelemetry-processor-baggage`
@@ -37,8 +40,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#4438](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4438))
 
 ### Fixed
-- `opentelemetry-instrumentation`: Batch all packages into a single `pip install` call in `bootstrap.py` to improve performance
-  ([#4530](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4530))
 - `opentelemetry-instrumentation-pika` Use `ObjectProxy` instead of `BaseObjectProxy` for `ReadyMessagesDequeProxy` to restore iterability with wrapt 2.x
   ([#4461](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4461))
 - `opentelemetry-instrumentation-dbapi` Use `ObjectProxy` instead of `BaseObjectProxy` for `TracedCursorProxy` to restore iterability with wrapt 2.x